Types: Add UA_DataTypeArray as a linked list for client/server

+++ src/ua_types.c
@@ -1086,9 +1086,9 @@ UA_Array_delete(void *p, size_t size, const UA_DataType *type) {
 
 UA_Boolean
 isDataTypeNumeric(const UA_DataType *type) {
-    // All data types ids between UA_TYPES_SBYTE and UA_TYPES_DOUBLE are numeric
-    for (int i = UA_TYPES_SBYTE; i <= UA_TYPES_DOUBLE; ++i)
-        if (&UA_TYPES[i] == type)
+    /* All data types between UA_TYPES_SBYTE and UA_TYPES_DOUBLE are numeric */
+    for(size_t i = UA_TYPES_SBYTE; i <= UA_TYPES_DOUBLE; ++i)
+        if(&UA_TYPES[i] == type)
             return true;
     return false;
 }
